# src/integration-test/resources/openapi.yml
openapi: 3.0.1
info:
  title: Messaging
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "3.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /status/{messageId}:
    get:
      tags:
      - Status and History Resources
      summary: Get the status for a single message and its deliveries
      operationId: getMessageStatus
      parameters:
      - name: messageId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResult'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/MessageResult'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /message/{messageId}:
    get:
      tags:
      - Status and History Resources
      summary: Get a message and all its deliveries
      operationId: getMessage
      parameters:
      - name: messageId
        in: path
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /conversation-history/{partyId}:
    get:
      tags:
      - Status and History Resources
      summary: Get the entire conversation history for a given party
      operationId: getConversationHistory
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
      - name: from
        in: query
        description: "From-date (inclusive). Format: yyyy-MM-dd (ISO8601)"
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: "To-date (inclusive). Format: yyyy-MM-dd (ISO8601)"
        required: false
        schema:
          type: string
          format: date
      responses:
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
  /batch-status/{batchId}:
    get:
      tags:
      - Status and History Resources
      summary: "Get the status for a message batch, its messages and their deliveries"
      operationId: getBatchStatus
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBatchResult'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/MessageBatchResult'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    MessageResult:
      type: object
      properties:
        messageId:
          type: string
          description: The message id
          format: uuid
        deliveryId:
          type: string
          description: The delivery id
          format: uuid
          nullable: true
        status:
          $ref: '#/components/schemas/MessageStatus'
      description: Message result
    MessageStatus:
      type: string
      description: Status
      enum:
      - PENDING
      - AWAITING_FEEDBACK
      - SENT
      - NOT_SENT
      - FAILED
      - NO_FEEDBACK_SETTINGS_FOUND
      - NO_FEEDBACK_WANTED
...
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        title:
          type: string
        detail:
          type: string
        status:
          $ref: '#/components/schemas/StatusType'
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
...
    MessageBatchResult:
      type: object
      properties:
        batchId:
          type: string
          description: The batch id
          format: uuid
        messages:
          type: array
          description: The individual message results
          items:
            $ref: '#/components/schemas/MessageResult'
      description: Message batch result
...
    HistoryResponse:
      type: object
      properties:
        messageType:
          type: string
          enum:
          - MESSAGE
          - EMAIL
          - SMS
          - WEB_MESSAGE
          - DIGITAL_MAIL
          - SNAIL_MAIL
          - LETTER
        status:
          type: string
          enum:
          - PENDING
          - AWAITING_FEEDBACK
          - SENT
          - NOT_SENT
          - FAILED
          - NO_FEEDBACK_SETTINGS_FOUND
          - NO_FEEDBACK_WANTED
        content:
          type: object
        timestamp:
          type: string
          format: date-time
  securitySchemes: {}